+++ src/mol-model-formats/shape/ply.ts
@@ -21,6 +21,7 @@ import { deepClone } from '../../mol-util/object';
 import { stringToWords } from '../../mol-util/string';
 import { ValueCell } from '../../mol-util/value-cell';
 import { Mat4 } from '../../mol-math/linear-algebra/3d/mat4';
+import { ShapeWireframeParams, wireframeShapeGetter } from './common';
 
 // TODO support 'edge' element, see https://www.mathworks.com/help/vision/ug/the-ply-format.html
 // TODO support missing face element
@@ -78,6 +79,7 @@ function createPlyShapeParams(plyFile?: PlyFile) {
 
     return {
         ...Mesh.Params,
+        ...ShapeWireframeParams,
 
         coloring: PD.MappedStatic(defaultColoring, {
             vertex: PD.Group({
@@ -300,12 +302,14 @@ function makeShapeGetter() {
 
 export function shapeFromPly(source: PlyFile, params?: { transforms?: Mat4[] }) {
     return Task.create<ShapeProvider<PlyData, Mesh, PlyShapeParams>>('Shape Provider', async ctx => {
+        const getShape = makeShapeGetter();
         return {
             label: 'Mesh',
             data: { source, transforms: params?.transforms },
             params: createPlyShapeParams(source),
-            getShape: makeShapeGetter(),
-            geometryUtils: Mesh.Utils
+            getShape,
+            geometryUtils: Mesh.Utils,
+            getWireframeShape: wireframeShapeGetter(getShape)
         };
     });
 }